Kokkos: addding a section about the kokkos programming model
The section is adding in the data services general section.
It currently focuses on the technical details of the Kokkos programming
model itself. An additional section on the motivation for the Kokkos
ecosystem, its impact on other Trilinos packages  and other aspects such
as the HPSF can be added in the introduction as suggested by Michael.

Signed-off-by: Luc Berger-Vergiat <lberge@sandia.gov>